requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
sqlalchemy>=2.0.0
passlib>=1.7.0
python-jose[cryptography]>=3.3.0
python-dotenv